fix: Profile 등록 로직 수정
- ImageUploader 위치 이동: 외부기술(S3)에 의존하는 서비스의 인터페이스 위치를 DDD 전략에 따라 이동
- yml 프로퍼티 주입: .yml 파일에 위치한 프로퍼티를 Java 타입으로 관리하도록 프로퍼티 바인딩 클래스 생성
- 이미지 업로드 메소드 변경: deprecated 된 메소드 create() -> createFrom()
- 이미지 업로드 확장자 변경: Files.probeContentType() 활용
- 클래스 위치 이동: 기능별로 구성(이미지는 프로필에만 사용되기 때문에 프로필 이하로 이동)
