+++ src/main/java/com/avalon/avalonchat/domain/profile/repository/BackgroundImageRepository.java
@@ -1,6 +1,6 @@
-package com.avalon.avalonchat.infra.image.upload;
+package com.avalon.avalonchat.domain.profile.repository;
 
-import com.avalon.avalonchat.domain.profile.domain.image.BackgroundImage;
+import com.avalon.avalonchat.domain.profile.domain.BackgroundImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BackgroundImageRepository extends JpaRepository<BackgroundImage, Long> {
