+++ src/main/java/com/avalon/avalonchat/domain/profile/repository/ProfileImageRepository.java
@@ -1,6 +1,6 @@
-package com.avalon.avalonchat.infra.image.upload;
+package com.avalon.avalonchat.domain.profile.repository;
 
-import com.avalon.avalonchat.domain.profile.domain.image.ProfileImage;
+import com.avalon.avalonchat.domain.profile.domain.ProfileImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfileImageRepository extends JpaRepository<ProfileImage, Long> {
